Rename hook to convey rich rather than remote
